+++ clients/pipeline-service/pms-sdk-core/src/test/java/io/harness/pms/sdk/core/execution/invokers/AsyncStrategyTest.java
@@ -104,7 +104,7 @@ public class AsyncStrategyTest extends PmsSdkCoreTestBase {
 
     Mockito.verify(asyncWaitEngine, Mockito.times(1))
         .waitForAllOn(notifyCallbackArgumentCaptor.capture(), progressCallbackArgumentCaptor.capture(),
-            correlationIdsCaptor.capture(), eq(TestAsyncStep.timeout));
+            correlationIdsCaptor.capture(), eq((long) TestAsyncStep.timeout));
     // Wait Engine Mock verify interactions
     AsyncSdkResumeCallback resumeCallback = notifyCallbackArgumentCaptor.getValue();
     assertThat(resumeCallback.getAmbianceBytes()).isEqualTo(ambiance.toByteArray());
